[homework-6-spring-orm]
[homework-6-spring-orm]

+++ src/main/java/com/github/kotelkov/pms/service/UserProfileService.java
@@ -1,13 +1,13 @@
 package com.github.kotelkov.pms.service;
 
-import com.github.kotelkov.pms.model.UserProfile;
+import com.github.kotelkov.pms.entity.UserProfile;
 
 import java.util.List;
 
 public interface UserProfileService {
     void createUserProfile(UserProfile userProfile);
-    UserProfile getUserProfileById(int id);
+    UserProfile getUserProfileById(Long id);
     List<UserProfile> getAllUsersProfiles();
     boolean updateUserProfile(UserProfile userProfile);
-    boolean deleteUserProfileById(int id);
+    boolean deleteUserProfileById(Long id);
 }
